Add game constructors. Add error checking to controllers.

+++ src/main/java/com/lukecahill/spring/controllers/GameController.java
@@ -1,5 +1,7 @@
 package com.lukecahill.spring.controllers;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.lukecahill.spring.models.Game;
 import com.lukecahill.spring.services.GameService;
 import com.lukecahill.spring.viewmodels.GameViewModel;
@@ -15,6 +17,8 @@ import javax.inject.Inject;
 public class GameController {
 
     private GameService gameService;
+    private GsonBuilder builder = new GsonBuilder();
+    private Gson gson = builder.create();
 
     public GameController() {}
 
@@ -24,7 +28,7 @@ public class GameController {
     }
 
     @RequestMapping(method = RequestMethod.GET, path = "/{gameId}")
-    public ResponseEntity<?> get(@PathVariable int gameId, Errors errors) {
+    public ResponseEntity<?> get(@PathVariable int gameId) {
         return ResponseEntity.ok(gameService.get(gameId));
     }
 
@@ -35,17 +39,29 @@ public class GameController {
 
     @RequestMapping(method = RequestMethod.POST, path = "/{gameId}")
     public ResponseEntity<?> add(@PathVariable int gameId, Game game, Errors errors) {
+        if(errors.hasErrors()) {
+            return ResponseEntity.badRequest().body(errors.getAllErrors());
+        }
+
         GameViewModel gameViewModel = gameService.add(game);
         return ResponseEntity.ok(gameViewModel);
     }
 
     @RequestMapping(method = RequestMethod.PUT, path = "/{gameId}")
     public ResponseEntity<?> update(@PathVariable int gameId, Game game, Errors errors) {
+        if(errors.hasErrors()) {
+            return ResponseEntity.badRequest().body(errors.getAllErrors());
+        }
+
         return ResponseEntity.ok(gameService.update(gameId, game));
     }
 
     @RequestMapping(method = RequestMethod.DELETE, path = "/{gameId}")
     public ResponseEntity<?> delete(@PathVariable int gameId, Errors errors) {
+        if(errors.hasErrors()) {
+            return ResponseEntity.badRequest().body(errors.getAllErrors());
+        }
+
         gameService.delete(gameId);
         return ResponseEntity.ok().build();
     }
